[#S05P12A602-479] Fix : DTO Long type 변수 -> int type 변경

+++ commbServer/src/main/java/com/ssafy/commb/dto/event/DailyEventDto.java
@@ -17,14 +17,14 @@ public class DailyEventDto {
     private KeywordDto keyword;
 
     static class Daily{
-        Long id;
+        int id;
     }
 
     @Getter
     @AllArgsConstructor
     @NoArgsConstructor
     public static class Request{                // 요청
-        private Long today;
+        private int today;
     }
 
     @Getter
